Trait text_writer::TextWriter

pub trait TextWriter {
    fn write_str(&mut self, s: &str) -> Result;

    fn write_char(&mut self, c: char) -> Result { ... }
    fn write_fmt(&mut self, args: Arguments) -> Result { ... }
}

A Unicode write-only stream.

Required Methods

fn write_str(&mut self, s: &str) -> Result

Write a string.

A mininmal implementation can have only this method.

Provided Methods

fn write_char(&mut self, c: char) -> Result

Write a code point.

A default implementation based on write_str is provided, but it is recommended to override it if it can be done more efficiently.

fn write_fmt(&mut self, args: Arguments) -> Result

Make TextWriter usable with the write! macro.

This typically should not be overridden